requirements.txt
requests
pandas
xarray
netCDF4
python-dotenv
pyarrow
duckdb
polars
tqdm
aiohttp
aiofiles
tenacity